requirements.txt
google-generativeai
openpyxl
python-dotenv
boto3